requirements.txt
Django==1.11.18
celery==4.2.1
redis==3.0.1
djangorestframework==3.9.0
